Refactored the code to have a common place where I compute a bunch of things

+++ src/raytracer/lighting/CheckersPattern.cpp
@@ -5,7 +5,7 @@ CheckersPattern::CheckersPattern(const Color &first, const Color &second) {
     m_second = second;
 }
 
-Color CheckersPattern::colorAt(const glm::vec3 &point) const {
+Color CheckersPattern::colorAt(const glm::dvec3 &point) const {
     auto sum = static_cast<int>(floor(point.x) + floor(point.y) + floor(point.z));
     if (sum % 2 == 0) {
         return m_first;
