+++ src/raytracer/lighting/GradientPattern.h
@@ -10,7 +10,7 @@ public:
     GradientPattern(const GradientPattern &other);
     ~GradientPattern() override = default;
 
-    [[nodiscard]] Color colorAt(const glm::vec3 &point) const override;
+    [[nodiscard]] Color colorAt(const glm::dvec3 &point) const override;
 private:
     Color m_first { Color::black() }, m_second { Color::white() };
 };
