+++ src/raytracer/lighting/StripePattern.h
@@ -10,7 +10,7 @@ public:
     StripePattern(const StripePattern& other);
     ~StripePattern() override = default;
 
-    [[nodiscard]] Color colorAt(const glm::vec3 &point) const override;
+    [[nodiscard]] Color colorAt(const glm::dvec3 &point) const override;
 private:
     Color m_first { Color::black() }, m_second { Color::white() };
 };
